Split calculators from configuration management
