+++ src/mediapipe_calculators/modelapiovmsadapterwrapper.cc
@@ -13,14 +13,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //*****************************************************************************
-#include "modelapiovmsadapterwrapper.h"
+#include "modelapiovmsadapterwrapper.hpp"
 
 #include <memory>
 
 #include "mediapipe/framework/calculator_framework.h"
 #include "mediapipe/framework/port/canonical_errors.h"
-#include "modelapiovmsadapter.h"
-#include "src/mediapipe_internal/ovmscalculator.pb.h"
+#include "modelapiovmsadapter.hpp"
+#include "src/mediapipe_calculators/ovmscalculator.pb.h"
 // here we need to decide if we have several calculators (1 for OVMS repository, 1-N inside mediapipe)
 // for the one inside OVMS repo it makes sense to reuse code from ovms lib
 namespace mediapipe {
