+++ src/mediapipe_calculators/modelapiovmsinferencecalculator.cc
@@ -23,9 +23,9 @@
 #include "../stringutils.hpp"  // TODO dispose
 #include "mediapipe/framework/calculator_framework.h"
 #include "mediapipe/framework/port/canonical_errors.h"
-#include "modelapiovmsadapter.h"
-#include "modelapiovmsadapterwrapper.h"
-#include "src/mediapipe_internal/ovmscalculator.pb.h"
+#include "modelapiovmsadapter.hpp"
+#include "modelapiovmsadapterwrapper.hpp"
+#include "src/mediapipe_calculators/ovmscalculator.pb.h"
 // here we need to decide if we have several calculators (1 for OVMS repository, 1-N inside mediapipe)
 // for the one inside OVMS repo it makes sense to reuse code from ovms lib
 namespace mediapipe {
