+++ src/mediapipe_calculators/ovms_tf_calculator.cc
